requirements.txt
langchain-openai>=0.0.5
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
tabulate>=0.9.0
python-dotenv>=0.19.0
colorama>=0.4.6
tqdm
rich
